chore: Update gojob and uio dependencies to latest versions
This commit updates the go.mod file to use the latest versions of the gojob and uio dependencies. The gojob dependency is updated to version v0.0.11-0.20240905152900-36da724ca9ae, and the uio dependency is updated to version v0.0.0-20240905152743-19ba2df5e6a6. These updates include bug fixes and performance improvements.